RUMM-301 Shopist project created
