+++ Shopist/Shopist.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
