+++ Shopist/Shopist/Info.plist
