+++ Shopist/Shopist/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
