+++ Shopist/Shopist/Resources/Assets.xcassets/Contents.json
