+++ Shopist/Shopist/Shopist.xcconfig
@@ -1,0 +1,1 @@
+#include? "shopist-secrets.local.xcconfig"
