+++ Shopist/Shopist/Utils/ConsoleOutputInterceptor.swift
