+++ Shopist/ShopistUITests/Info.plist
