--- examples/.gitignore
@@ -1,2 +1,0 @@
-# `examples-secret.xcconfig` file stores your Datadog secrets and should not be checked in
-examples-secret.xcconfig
