--- examples/iOS-app-example-cocoapods/iOS-app-example-cocoapods-UITests/.where-are-the-source-files
@@ -1,1 +1,0 @@
-Go to `../../examples-shared-uitests/` - example projects share UITests files through this folder.
