--- examples/iOS-app-example-cocoapods/iOS-app-example-cocoapods/.where-are-the-source-files
@@ -1,1 +1,0 @@
-Go to `../../examples-shared-sources/` - example projects share files through this folder.
