--- examples/iOS-app-example-objc/.gitignore
@@ -1,7 +1,0 @@
-xcuserdata/
-build/
-Pods/
-
-# Ignore .lock file as dependencies are linked using relative path reference
-# so they are never locked.
-Podfile.lock
