--- examples/iOS-app-example-objc/iOS-app-example-objc.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "self:iOS-app-example-objc.xcodeproj">
-   </FileRef>
-</Workspace>
