--- examples/iOS-app-example-objc/iOS-app-example-objc/ViewController.h
@@ -1,11 +1,0 @@
-/*
-* Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
-* This product includes software developed at Datadog (https://www.datadoghq.com/).
-* Copyright 2019-2020 Datadog, Inc.
-*/
-
-#import <UIKit/UIKit.h>
-
-@interface ViewController : UIViewController
-
-@end
