--- examples/iOS-app-example-spm/.gitignore
@@ -1,5 +1,0 @@
-xcuserdata/
-build/
-
-# Ignore `Package.resolved` file to always depend on recent `master`
-Package.resolved
