--- examples/iOS-app-example-spm/example-specific-config.xcconfig
@@ -1,5 +1,0 @@
-// The `examples-secret.xcconfig` is git-ignored thus it is included optionaly.
-// See `README.md` for more details about `examples-secret.xcconfig`.
-#include? "../examples-secret.xcconfig"
-
-EXAMPLE_APP_NAME=SPM-example
