--- examples/iOS-app-example-spm/iOS-app-example-spm-UITests/.where-are-the-source-files
@@ -1,1 +1,0 @@
-Go to `../../examples-shared-uitests/` - example projects share UITests files through this folder.
