--- examples/iOS-app-example-spm/iOS-app-example-spm/.where-are-the-source-files
@@ -1,1 +1,0 @@
-Go to `../../examples-shared-sources/` - example projects share files through this folder.
